fix(PhoneVerificationTokenService): fix TokenNotFound bug
Initially, we were removing the token from the cache at the end of this <see cref="ValidateTokenAsync"/> method. But, for the new user registrations, since we are requesting token twice; - First in <see cref="UseCases.Users.SelfRegister.Endpoint> - Then in <see cref="Tokens.UseCases.Create.Endpoint> If we remove the token from the cache after the first request, the second request will fail with <see cref="PhoneVerificationTokenErrors.TokenNotFound"/> And users will have to go back to the very first step of the registration process.

BREAKING CHANGE: